Make panda udev rules less likely to be overridden by another rules set
I spent about 3-4 hours diagnosing why Linux Mint 18.3 (based on Ubuntu 16.04) and panda wouldn't communicate properly. Found that the rules list in /lib/udev/rules.d/50-udev-defaults.rules overrides the udev rules we add for panda in /etc/udev/rules.d/11-panda.rules. Defaults set the device mode to 0664, while we want 0666. If we set the order for our rules to be read later/last , we have less of a chance of them being overridden and causing libusb1 to not be able to gain device access.