Update safety_toyota.h
Tss2 limit is above 15. Allows code to be edited on shanes SA. My indi tune with this is the best and i need to gather 5 routes to submit full tune PR. This is the first step. Does not effect vehicles just allows edits for routes for PR. Improved handing for corolla tss2 tune.